make configurable spyglass layers
make configurable spyglass layers by adding
configOptions.nonSpyGlassGraphcLayerIds which is an array of strings
representing layer ids that are NOT to be limited to the spy glass.
Without any values in this array, all graphic layers are only visible in
the spy glass.  Layer Ids included in
configOptions.nonSpyGlassGraphicLayerIds are also visible outside of the
spyglass.